Apply workaround to NodesHub limitation
As stated in issue #15, NodesHub is not able to correctly manage
spontaneous connection attempts.

This patch tells the nodes to not attempt performing these spontaneous
connections, so it does not fix the bug, but at least mitigates some of
its effects.

In addition to that, now we are more explicit about some node settings.

Signed-off-by: Andres Correa Casablanca <andres@thirdhash.com>